Crate mlx_ethernet

This crate defines the layout of memory objects that make up the software interface between the Mellanox hardware and the driver, as well as functions to access different fields of these objects.

The Mellanox ethernet card is referred to as both the NIC (Network Interface Card) and the HCA (Host Channel Adapter).

All information is taken from the Mellanox Adapters Programmer’s Reference Manual (PRM) [Rev 0.54], unless otherwise specified. While this version of the manual was acquired by directly contacting Nvidia through their support site (https://support.mellanox.com/s/), an older version of the manual can be found at http://www.mellanox.com/related-docs/user_manuals/Ethernet_Adapters_Programming_Manual.pdf.

Modules

command_queue

Defines the Command Queue that is used to pass commands from the driver to the NIC.

Structs

InitializationSegment

The initialization segment is located at offset 0 of PCI BAR0. It is used in the initialization procedure of the device, and it contains the 32-bit command doorbell vector used to inform the HW when a command is ready to be processed.

Enums

InitializingState

The possible values of the initialization state of the device as taken from the intialization segment.