Module mlx_ethernet::command_queue

Defines the Command Queue that is used to pass commands from the driver to the NIC.

Structs

CommandQueue
A buffer of fixed-size entries that is used to pass commands to the HCA. It resides in a physically contiguous 4 KiB memory chunk. (Section 8.24.1: HCA Command Queue)

CommandQueueEntry
Layout of a command passed to the NIC. The fields include control information for the command as well as actual command input and output. The first 16 bytes of the actual command input are part of the entry. The remaining data is written in mailboxes. Similarly, the first 16 bytes of the command output are part of the entry and remaining data is written in mailboxes.

Enums

CommandDeliveryStatus
Return codes written by HW in the delivery status field of the command entry. See CommandQueueEntry::token_signature_status_own.

CommandOpcode
Command opcode written by SW in opcode field of the input data in the command entry. See CommandQueueEntry::command_input_opcode.

CommandQueueError
Possible reasons for failure when executing a command

CommandReturnStatus
Command status written by HW in status field of the output data in the command entry. See CommandQueueEntry::command_output_status.

CommandTransportType
Type of transport that carries the command.

ManagePagesOpMod
Possible values of the opcode modifer when the opcode is CommandOpcode::ManagePages.

QueryPagesOpMod
Possible values of the opcode modifer when the opcode is CommandOpcode::QueryPages.