Enum mlx_ethernet::command_queue::CommandQueueError

pub enum CommandQueueError {
    NoCommandEntryAvailable,
    MissingInputPages,
    MissingInput,
    IncorrectCommandOpcode,
    InvalidCommandOpcode,
    InvalidCommandDeliveryStatus,
    InvalidCommandReturnStatus,
    CommandNotCompleted,
    InvalidMailboxOffset,
    PageAllocationFailed,
    UnimplementedOpcode,
    NotImplemented,
}

Possible reasons for failure when executing a command

Variants

NoCommandEntryAvailable

All command entries are currently being used

MissingInputPages

Allocated pages are not passed to a command that requires them

MissingInput

Any other input is not passed to a command that requires them

IncorrectCommandOpcode

Opcode value in the command entry is not what was expected

InvalidCommandOpcode

Opcode in the command entry is not a valid value

InvalidCommandDeliveryStatus

Delivery status in the command entry is not a valid value

InvalidCommandReturnStatus

Return status in the command entry is not a valid value

CommandNotCompleted

Trying to access the command entry before HW is done processing it

InvalidMailboxOffset

Offset in a page is too large to map a [CommandInterfaceMailbox] to that offset

PageAllocationFailed

A call to create a MappedPages failed

UnimplementedOpcode

Initializing a comand entry for the given opcode has not been implemented

NotImplemented

Some function has not been implemented for the given opcode