Struct mlx_ethernet::command_queue::CommandQueue

#[repr(C)]
pub struct CommandQueue { /* fields omitted */ }

A buffer of fixed-size entries that is used to pass commands to the HCA. It resides in a physically contiguous 4 KiB memory chunk. (Section 8.24.1: HCA Command Queue)

Implementations

impl CommandQueue

pub fn create(
entries: BoxRefMut<MappedPages, [CommandQueueEntry]>,
num_cmdq_entries: usize
) -> Result<CommandQueue, &'static str>

Create a command queue object.

Arguments

entries: physically contiguous memory that is mapped as a slice of command queue entries.

num_cmdq_entries: number of entries in the queue.

pub fn create_command(
&mut self,
opcode: CommandOpcode,
opmod: Option<u16>,
allocated_pages: Option<Vec<PhysicalAddress>>
) -> Result<usize, CommandQueueError>

Fill in the fields of a command queue entry. At the end of the function, the command is ready to be posted using the doorbell in the initialization segment. Returns an error if no entry is available to use.

Arguments

opcode: CommandOpcode for the command that the driver wants to execute

opmod: opcode modifer, only applicable for certain commands

allocated_pages: physical address of pages that need to be passed to the NIC. Only used in the CommandOpcode::ManagePages command.

pub fn wait_for_command_completion(
&mut self,
entry_num: usize
) -> Result<(CommandDeliveryStatus, CommandReturnStatus), CommandQueueError>

Waits for ownership bit to be cleared, and then returns the command delivery status and the command return status.

pub fn get_query_issi_command_output(
&self,
entry_num: usize
) -> Result<(u16, u8), CommandQueueError>

Get the current ISSI version and the supported ISSI versions, which is the output of the CommandOpcode::QueryIssi command.

pub fn get_query_pages_command_output(
&self,
entry_num: usize
) -> Result<u32, CommandQueueError>

Get the number of pages requested by the NIC, which is the output of the CommandOpcode::QueryPages command.

pub fn get_uar(&self, entry_num: usize) -> Result<u32, CommandQueueError>

Get the User Access Region (UAR) number, which is the output of the CommandOpcode::AllocUar command.

pub fn get_protection_domain(
&self,
entry_num: usize
) -> Result<u32, CommandQueueError>

Get the protection domain number, which is the output of the CommandOpcode::AllocPd command.

pub fn get_transport_domain(
&self,
entry_num: usize
) -> Result<u32, CommandQueueError>

Get the transport domain number, which is the output of the CommandOpcode::AllocTransportDomain command.

pub fn get_reserved_lkey(
&self,
entry_num: usize
) -> Result<u32, CommandQueueError>

Get the value of the reserved Lkey for Base Memory Management Extension, which is used when we are using physical addresses. It is taken as the output of the CommandOpcode::QuerySpecialContexts command.