Module mlx_ethernet::completion_queue

Completion Queues (CQ) are circular buffers used by the HCA to post completion reports upon completion of a work request. This module defines the layout of an CQ, the context used to initialize an CQ and related functions.

(PRM Section 8.18: Completion Queues)

Structs

CompletionQueue

A data structure that contains the CQ buffer and is used to interact with the CQ once initialized.

CompletionQueueDoorbellRecord

A structure containing information of recently-posted CQ commands

CompletionQueueEntry

The layout of an entry in the CQ buffer.