Struct mlx_ethernet::initialization_segment::InitializationSegment

#[repr(packed)]
pub struct InitializationSegment { /* private fields */ }

The initialization segment is located at offset 0 of PCI BAR0. It is used in the initialization procedure of the device, and it contains the 32-bit command doorbell vector used to inform the HW when a command is ready to be processed.

Implementations

impl InitializationSegment

pub fn num_cmdq_entries(&self) -> u8

Returns the maximum number of entries that can be in the command queue

pub fn cmdq_entry_stride(&self) -> u8

Returns the required stride of command queue entries (bytes between the start of consecutive entries)

pub fn set_physical_address_of_cmdq(
&mut self,
cmdq_physical_addr: PhysicalAddress
) -> Result<(), &'static str>

Sets the physical address of the command queue within the initialization segment.

Arguments

cmdq_physical_addr: the starting physical address of the command queue, the lower 12 bits of which must be zero.

pub fn device_is_initializing(&self) -> bool

Returns true if the device is still initializing, and driver should not pass any commands to the device.