Struct mlx_ethernet::receive_queue::ReceiveQueue

pub struct ReceiveQueue { /* private fields */ }

A data structure that contains the RQ ring of descriptors and is used to interact with the RQ once initialized.

Implementations

impl ReceiveQueue

pub fn create(
entries_mp: MappedPages,
num_entries: usize,
mtu: u32,
pool: &'static Queue<ReceiveBuffer>,
rqn: Rqn,
lkey: Lkey,
cq: CompletionQueue
) -> Result<ReceiveQueue, &'static str>

Creates a RQ by mapping the buffer as a slice of WorkQueueEntryReceives. Each WQE is set to an initial state.

Arguments

entries_mp: memory that is to be transformed into a slice of WQEs. The starting physical address should have been passed to the HCA when creating the SQ.

num_entries: number of entries in the RQ

mtu: size of the receive buffers in bytes

buffer_pool: receive buffer pool

rqn: SQ number returned by the HCA

lkey: the lkey used by the RQ

pub fn refill(&mut self) -> Result<(), &'static str>

Refills the receive queue by updating WQEs with new packet buffers. Right now we assume that this function is only called once at the point of initialization.

TODO: this function can be shifted to nic_initialization if we remove intel specific actions from those functions