Module mlx_ethernet::send_queue

The Send Queue (SQ) object holds the descriptor ring used to send outgoing messages and packets. The descriptor ring is referred to as a Work Queue Buffer. This module defines the layout of an SQ, the context used to initialize a SQ, the Transport Interface Send object attached to the queue and related functions.

(PRM Section 8.15: Send Queue)

Structs

SendQueue

A data structure that contains the SQ ring of descriptors and is used to interact with the SQ once initialized.

Enums

SendQueueState

The possible states the SQ can be in.